Changed folder name to lwer case
