+++ src/components/cart-dropdown/CartDropdown.scss
