+++ src/components/cart-icon/CartIcon.jsx
