+++ src/components/cart-icon/CartIcon.scss
