+++ src/components/cart-item/CartItem.jsx
