+++ src/components/cart-item/CartItem.scss
