+++ src/components/category-item/CategoryItem.scss
