+++ src/components/category-preview/CategoryPreview.scss
