+++ src/components/checkout-item/CheckoutItem.scss
