+++ src/components/login-form/LoginForm.jsx
@@ -1,52 +1,45 @@
-import "./RegisterForm.scss";
-import { useState } from "react";
-import {
-    signInWithGooglePopup,
-    createAuthUserWithEmailAndPassword,
-    createUserDocFromAuth,
-} from "../../utils/firebase/firebase";
-import FormInput from "../FormInput/FormInput";
-import { defaultRegisterFormFields, RegisterFormInputData } from "../../database/RegisterFormInputData";
-import Button from "../Button/Button";
+import './LoginForm.scss'
+import FormInput from "../form-input/FormInput";
+import Button from "../button/Button";
 import GoogleLogo from "../../assets/google-logo.png";
 import { Link, useNavigate } from "react-router-dom";
+import { defaultLoginFormFields, LoginFormInputData } from "../../database/login-form-input-data";
+import { useState } from 'react';
+import { loginAuthUserWithEmailAndPassword } from '../../utils/firebase/firebase'
+import {
+    signInWithGooglePopup,
+} from "../../utils/firebase/firebase";
 
-const RegisterForm = () => {
-    const [formFields, setFormFields] = useState(defaultRegisterFormFields);
-    const { displayName, email, password, confirmPassword } = formFields;
+const LoginForm = () => {
+    const [formFields, setFormFields] = useState(defaultLoginFormFields);
+    const { email, password } = formFields;
     const [error, setError] = useState('');
-    const [regMsg, setRegMsg] = useState('');
+    const [loggedInMsg, setLoggedInMsg] = useState('');
     const navigate = useNavigate();
 
     const logGooglePopupUser = async () => {
         await signInWithGooglePopup();
-        setRegMsg("Successfully registered");
+        setLoggedInMsg("Successfully logged in");
         setTimeout(() => { navigate('/') }, 2000);
     };
 
     const resetFormFields = () => {
-        setFormFields(defaultRegisterFormFields);
+        setFormFields(defaultLoginFormFields);
     }
 
     const handleSubmit = async (e) => {
         e.preventDefault();
 
-        if (password !== confirmPassword) {
-            setError("Passwords do not match");
-            return;
-        }
-
         try {
-            const { user } = await createAuthUserWithEmailAndPassword(email, password);
-            await createUserDocFromAuth(user, { displayName });
+            await loginAuthUserWithEmailAndPassword(email, password);
+            setLoggedInMsg("Successfully logged in");
             resetFormFields();
-            setRegMsg("Successfully created account");
             setTimeout(() => { navigate('/') }, 2000);
         } catch (error) {
-            if (error.code === 'auth/email-already-in-use') {
-                setError('Email already in use')
-            } else if (error.code === 'auth/weak-password') {
-                setError('Password is too weak')
+            if (error.code === 'auth/user-not-found') {
+                setError('User not found')
+            } else if (error.code === 'auth/wrong-password') {
+                setError('Wrong password')
             } else {
                 setError(error.message)
             }
@@ -59,15 +52,14 @@ const RegisterForm = () => {
     };
 
     return (
-        <div className="register-form-container">
+        <div className="login-form-container">
             <h2 className="register-title">
-                Create your account
+                Login to your account
                 <br />
-                to get started
+                to buy awesome clothes
             </h2>
-
             <form onSubmit={handleSubmit}>
-                {RegisterFormInputData.map((inputData) => {
+                {LoginFormInputData.map((inputData) => {
                     const { label, required, type, name } = inputData;
 
                     return (
@@ -82,17 +74,17 @@ const RegisterForm = () => {
                     )
                 })}
 
-                <span className='msg-reg'>
+                <span className='msg-logged-in'>
                     {
-                        regMsg ? (
-                            <p>{regMsg}</p>
+                        loggedInMsg ? (
+                            <p>{loggedInMsg}</p>
                         ) : (
                             <p></p>
                         )
                     }
                 </span>
 
-                <span className='error-reg'>
+                <span className='error-login'>
                     {
                         error ? (
                             <p>{error}</p>
@@ -106,7 +98,7 @@ const RegisterForm = () => {
                     btnType="primaryBtn"
                     type="submit"
                 >
-                    Register
+                    Login
                 </Button>
 
                 <Button
@@ -114,19 +106,20 @@ const RegisterForm = () => {
                     type="button"
                     onClick={logGooglePopupUser}
                 >
-                    <img src={GoogleLogo} alt='GOogle Logo' className='google-logo' />
-                    Register with Google
+                    <img src={GoogleLogo} alt='Google Logo' className='google-logo' />
+                    Login with Google
                 </Button>
 
-                <div className="have-account">
-                    Already have an account?
-                    <Link to="/login">
-                        <span className="login-link">Login</span>
+                <div className="no-account">
+                    Don't have an account?
+                    <Link to="/register">
+                        <span className="register-link">Register</span>
                     </Link>
                 </div>
+
             </form>
         </div>
-    );
-};
+    )
+}
 
-export default RegisterForm;
+export default LoginForm
