+++ src/routes/Shop/Shop.jsx
@@ -1,6 +1,6 @@
 import { Routes, Route } from 'react-router-dom';
-import CategoriesPreview from "../CategoriesPreview/CategoriesPreview";
-import Category from '../Category/Category';
+import CategoriesPreview from "../categories-preview/CategoriesPreview";
+import Category from '../category/Category';
 
 const Shop = () => {
