+++ src/routes/categories-preview/CategoriesPreview.jsx
@@ -1,7 +1,7 @@
 import "./CategoriesPreview.scss";
 import { useContext } from "react";
 import { CategoriesContext } from "../../context/categories.context";
-import CategoryPreview from "../../components/CategoryPreview/CategoryPreview";
+import CategoryPreview from "../../components/category-preview/CategoryPreview";
 import { ReactComponent as CrownLogo } from "../../assets/crown-logo.svg";
 
 const CategoriesPreview = () => {
