+++ src/routes/categories-preview/CategoriesPreview.scss
